add readme update

+++ sdk/formrecognizer/azure-ai-formrecognizer/src/samples/java/com/azure/ai/formrecognizer/FormRecognizerAsyncClientJavaDocCodeSnippets.java
@@ -404,8 +404,8 @@ public class FormRecognizerAsyncClientJavaDocCodeSnippets {
                                     Map<String, FormField<?>> formFieldMap = FieldValueType.MAP.cast(receiptItem);
                                     formFieldMap.forEach((key, formField) -> {
                                         if ("Quantity".equals(key)) {
-                                            if (FieldValueType.FLOAT.equals(formField.getValueType())) {
-                                                Float quantity = FieldValueType.FLOAT.cast(formField);
+                                            if (FieldValueType.DOUBLE.equals(formField.getValueType())) {
+                                                Float quantity = FieldValueType.DOUBLE.cast(formField);
                                                 System.out.printf("Quantity: %f, confidence: %.2f%n",
                                                     quantity, formField.getConfidence());
                                             }
@@ -486,7 +486,8 @@ public class FormRecognizerAsyncClientJavaDocCodeSnippets {
                                 });
                             }
                         }
-                    }}));
+                    }
+                }));
         // END: com.azure.ai.formrecognizer.FormRecognizerAsyncClient.beginRecognizeReceipts#Flux-long
     }
